Moving CLI tools to co.rsk.cli.tools package
