+++ rskj-core/src/main/java/co/rsk/cli/tools/ExportState.java
@@ -15,8 +15,9 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package co.rsk;
+package co.rsk.cli.tools;
 
+import co.rsk.RskContext;
 import co.rsk.trie.NodeReference;
 import co.rsk.trie.Trie;
 import co.rsk.trie.TrieStore;
@@ -28,14 +29,10 @@ import java.io.PrintStream;
 import java.util.Optional;
 
 /**
- * The entry point for show state info CLI util
+ * The entry point for export state CLI tool
  * This is an experimental/unsupported tool
  */
-public class ShowStateInfo {
-    private static int nnodes;
-    private static int nvalues;
-    private static int nbytes;
-
+public class ExportState {
     public static void main(String[] args) {
         RskContext ctx = new RskContext(args);
         BlockStore blockStore = ctx.getBlockStore();
@@ -43,38 +40,25 @@ public class ShowStateInfo {
 
         execute(args, blockStore, trieStore, System.out);
     }
-    
+
     public static void execute(String[] args, BlockStore blockStore, TrieStore trieStore, PrintStream writer) {
-        Block block;
+        long blockNumber = Long.parseLong(args[0]);
 
-        if ("best".equals(args[0])) {
-            block = blockStore.getBestBlock();
-        }
-        else {
-            block = blockStore.getChainBlockByNumber(Long.parseLong(args[0]));
-        }
-
-        writer.println("Block number: " + block.getNumber());
-        writer.println("Block hash: " + Hex.toHexString(block.getHash().getBytes()));
-        writer.println("Block parent hash: " + Hex.toHexString(block.getParentHash().getBytes()));
-        writer.println("Block root hash: " + Hex.toHexString(block.getStateRoot()));
+        Block block = blockStore.getChainBlockByNumber(blockNumber);
 
         Optional<Trie> otrie = trieStore.retrieve(block.getStateRoot());
 
-        if (otrie.isPresent()) {
-            Trie trie = otrie.get();
-
-            processTrie(trie);
+        if (!otrie.isPresent()) {
+            return;
         }
 
-        writer.println("Trie nodes: " + nnodes);
-        writer.println("Trie long values: " + nvalues);
-        writer.println("Trie MB: " + (double)nbytes / (1024*1024));
+        Trie trie = otrie.get();
+
+        processTrie(trie, writer);
     }
 
-    private static void processTrie(Trie trie) {
-        nnodes++;
-        nbytes += trie.getMessageLength();
+    private static void processTrie(Trie trie, PrintStream writer) {
+        writer.println(Hex.toHexString(trie.toMessage()));
 
         NodeReference leftReference = trie.getLeft();
 
@@ -85,12 +69,11 @@ public class ShowStateInfo {
                 Trie leftTrie = left.get();
 
                 if (!leftReference.isEmbeddable()) {
-                    processTrie(leftTrie);
+                    processTrie(leftTrie, writer);
                 }
 
                 if (leftTrie.hasLongValue()) {
-                    nvalues++;
-                    nbytes += leftTrie.getValue().length;
+                    writer.println(Hex.toHexString(leftTrie.getValue()));
                 }
             }
         }
@@ -104,19 +87,17 @@ public class ShowStateInfo {
                 Trie rightTrie = right.get();
 
                 if (!rightReference.isEmbeddable()) {
-                    processTrie(rightTrie);
+                    processTrie(rightTrie, writer);
                 }
 
                 if (rightTrie.hasLongValue()) {
-                    nvalues++;
-                    nbytes += rightTrie.getValue().length;
+                    writer.println(Hex.toHexString(rightTrie.getValue()));
                 }
             }
         }
 
         if (trie.hasLongValue()) {
-            nvalues++;
-            nbytes += trie.getValue().length;
+            writer.println(Hex.toHexString(trie.getValue()));
         }
     }
 }
