+++ rskj-core/src/test/java/co/rsk/cli/tools/CliToolsTest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package co.rsk;
+package co.rsk.cli.tools;
 
 import co.rsk.core.BlockDifficulty;
 import co.rsk.crypto.Keccak256;
